Allow the module for each puzzle to return a result that can be checked
This commit adds a --check command line argument which adds checking
that the correct result was returned by each puzzle script.  This also
includes printing of the result to the screen regardless of whether
--check is used.

This still allows modules to do their own printing of the result if
checking is not being done.