Improve performance of 2022 day 14 by 15x
If the current sand grain comes to a stop after moving diagonally to
the right, then the next grain is going to follow the same path and
be stopped by the current grain.  The same is true when moving to the
left if the path too the right is blocked.  From this, whole diagonal
lines can be filled in at once.