Allow --day argument value to be a string
The --day command line option can now accept a string as its value.
'day' will still be prepended to that when importing the module for that
day, and the intended usage is to allow arguments like "--day=7_part1".

The read_data function has also been updated to this, but it requires
that an underscore be used to separate the day number from anything
after it.